requirements.txt
streamlit
pandas
matplotlib
seaborn
langchain-groq
langchain-experimental
python-dotenv
tabulate